+++ src/direction.h
@@ -1,6 +1,0 @@
-#ifndef __DIRECTION_H__
-#define __DIRECTION_H__
-
-typedef enum DIRECTION { LEFT, RIGHT, UP, DOWN };
-
-#endif // __DIRECTION_H__