+++ src/functions_resize.h
@@ -1,9 +1,0 @@
-#ifndef __FONCTIONSRESIZE__
-#define __FONCTIONSRESIZE__
-
-#include "multimonitor_move.h"
-
-extern bool ResizeWindow(const int hotkey, bool fromKbd = true);
-extern void MoveToScreen(DIRECTION);
-
-#endif //__FONCTIONSRESIZE__