+++ src/hotkey_number.h
@@ -1,6 +1,0 @@
-#ifndef __HOTKEY_NUMBER_H__
-#define __HOTKEY_NUMBER_H__
-
-#define NB_HOTKEYS 17
-
-#endif // __HOTKEY_NUMBER_H__