+++ src/layout_screens.h
@@ -1,7 +1,0 @@
-#ifndef ECRANS_PNG_H_INCLUDED
-#define ECRANS_PNG_H_INCLUDED
-
-wxImage wxGetImage_Ecran_16x9();
-wxImage wxGetImage_Ecran_4x3();
-
-#endif // ECRANS_PNG_H_INCLUDED